BOT CLASS Updated the constructor, for readability and traceability purposes. The constructor now receives all the information from .json which allows it to set all the properties in advance. Enabled the placement of bots either in the initial position or in an overloaded position if the arguments are provided to the function Place.
Included several types of AI:
-Brownian: the one that was already in place;
-Follower: no path finding and no recasting; in the persistent version, once the monster sees you he continues to chase you even after he is out of the line of sight. In the non-persistent version, he stops.

PORTAL
Similar updates as in Bot.

MAP
The constructor contains all the variables that are used by the object afterwards.
The initial information from the .json is passed on to the portals and the bots.

JSON
They were updated to include the ai_type for the bots.